requirements.txt
numpy==1.18.1
pandas==1.3
Pillow==7.1.2
tqdm==4.46.0
